+++ src/main/java/com/BankingAPI/BankingAPI/Group1/model/dto/TransactionGETPOSTResponseDTO.java
@@ -2,5 +2,5 @@ package com.BankingAPI.BankingAPI.Group1.model.dto;
 
 import java.time.LocalDate;
 
-public record TransactionGETPOSTResponseDTO(String fromAccount, String toAccount, double amount, LocalDate date, int userId) {
+public record TransactionGETPOSTResponseDTO(String fromAccount, String toAccount, double amount, LocalDate date, Long userId) {
 }
